[schema change, group principal name] Dropping group column "group_princpal_id" altogether, renaming "identifier" as "principal_name" and adding unique and non-nullabel constraints; dropping "grp_grpprincipalid_idx" index; updating field "identifier" in all group descriptors. For more on rationale for this change see the recent bungeni-dev email with subject "principal name for user and group types".
Adding alembic script to rename/drop columns, and morph data from group.type to principal.type columns; also dropping legacy column user.titles. To update execute the following in the bungeni folder: 
    ./bin/alembic upgrade head 
NOTE: for this update to succeed each value of group.identifier in the db must be UNIQUE - this means you may have to manually tweak the data in your db prior to running this upgrade to ensure that each value is unique as well as valid as per the configuration/vocabularies of your site.

Cleaning out event-based contortions to set group_principal_id, that had to be set after a group has been saved to the db; deleting core.events.group_created handler.

Updating unittests for bungeni.models.